finishing CRUD redux-toolkit

+++ src/components/TableComponent.jsx
@@ -1,5 +1,6 @@
 import {
   faCircleInfo,
+  faMagnifyingGlass,
   faPenToSquare,
   faTrashCan,
   faUser,
@@ -16,14 +17,13 @@ import {
   Pagination,
   PaginationItem,
   PaginationLink,
-  Row,
-  Col,
 } from "reactstrap";
 
 import { useDispatch, useSelector } from "react-redux";
-import { getUsers } from "../redux/services/getUsers";
+import swal from "sweetalert";
 import { deleteUser } from "../redux/services/deleteUser";
 import { getDetailUser } from "../redux/services/getDetailUser";
+import { getUsers } from "../redux/services/getUsers";
 import { clearDetailUser } from "../redux/userSlice";
 
 function TableComponent() {
@@ -41,7 +41,23 @@ function TableComponent() {
   };
 
   const handleDeleteUser = (userId) => {
-    dispatch(deleteUser(userId));
+    // warning alert delete
+    swal({
+      title: "Are you sure?",
+      text: "Once deleted, you will not be able to recover this user data !!!",
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
+    }).then((willDelete) => {
+      if (willDelete) {
+        dispatch(deleteUser(userId));
+        swal("Your imaginary files have been deleted !!!", {
+          icon: "success",
+        });
+      } else {
+        swal("Your user data is safe !!!");
+      }
+    });
   };
 
   useEffect(() => {
@@ -101,97 +117,96 @@ function TableComponent() {
 
   return (
     <Container>
-      <InputGroup className="mt-5 mb-1 w-25 float-end">
+      <InputGroup className="d-flex justify-content-between mt-5 mb-3 gap-5">
+        <Link to={"/create"}>
+          <Button className="shadow-lg">
+            <FontAwesomeIcon icon={faUser} className="me-2" />
+            Create user
+          </Button>
+        </Link>
         <Input
+          className="rounded shadow-lg"
           placeholder="Search..."
           value={searchTerm}
           onChange={handleSearch}
         ></Input>
       </InputGroup>
-      <Row>
-        <Col>
-          <Link to={"/create"}>
-            <Button>
-              <FontAwesomeIcon icon={faUser} className="me-2" />
-              Create user
-            </Button>
-          </Link>
-        </Col>
-      </Row>
-      <Table striped bordered hover>
-        <thead>
-          <tr>
-            <th>Id</th>
-            <th>Full Name</th>
-            <th>Alamat</th>
-            <th>Umur</th>
-            <th>No Handphone</th>
-            <th>Action</th>
-          </tr>
-        </thead>
-        <tbody>
-          {paginatedUsers.map((item) => (
-            <tr key={item.id}>
-              <td>{item.id}</td>
-              <td>{item.name}</td>
-              <td>{item.alamat}</td>
-              <td>{item.umur}</td>
-              <td>{item.noHp}</td>
-              <td className="d-flex justify-content-between">
-                <Link to={`/detail/${item.id}`}>
-                  <Button
-                    className="bg-info-subtle text-black mr-1"
-                    onClick={() => handleDetailUser(item.id)}
-                  >
-                    <FontAwesomeIcon icon={faCircleInfo} className="me-2" />
-                    Detail
-                  </Button>
-                </Link>
-                <Link to={"/edit/:id"}>
-                  <Button className="bg-success">
-                    <FontAwesomeIcon icon={faPenToSquare} className="me-2" />
-                    Edit
-                  </Button>
-                </Link>
-                <Button
-                  className="bg-danger"
-                  onClick={() => handleDeleteUser(item.id)}
-                >
-                  <FontAwesomeIcon icon={faTrashCan} className="me-2" />
-                  Delete
-                </Button>
-              </td>
+      <div className="table-responsive rounded">
+        <Table striped bordered hover className="shadow-lg">
+          <thead>
+            <tr>
+              <th className="text-danger">Id</th>
+              <th className="text-success">Full Name</th>
+              <th className="text-warning-emphasis">Alamat</th>
+              <th className="text-primary">Umur</th>
+              <th className="text-black">No Handphone</th>
+              <th className="text-warning">Action</th>
             </tr>
-          ))}
-        </tbody>
-      </Table>
-      <Pagination className="float-end">
-        <PaginationItem disabled={currentPage === 1}>
-          <PaginationLink first onClick={() => handlePageChange(1)} />
-        </PaginationItem>
-        <PaginationItem disabled={currentPage === 1}>
-          <PaginationLink
-            previous
-            onClick={() => handlePageChange(currentPage - 1)}
-          />
-        </PaginationItem>
-        {[...Array(totalPages).keys()].map((page) => (
-          <PaginationItem key={page} active={currentPage === page + 1}>
-            <PaginationLink onClick={() => handlePageChange(page + 1)}>
-              {page + 1}
-            </PaginationLink>
+          </thead>
+          <tbody>
+            {paginatedUsers.map((item) => (
+              <tr key={item.id}>
+                <td className="text-danger">{item.id}</td>
+                <td className="text-success">{item.name}</td>
+                <td className="text-warning-emphasis">{item.alamat}</td>
+                <td className="text-primary">{item.umur}</td>
+                <td className="text-black">{item.noHp}</td>
+                <td className="d-flex justify-content-between">
+                  <Link to={`/detail/${item.id}`}>
+                    <Button
+                      className="bg-info-subtle text-black mr-1"
+                      onClick={() => handleDetailUser(item.id)}
+                    >
+                      <FontAwesomeIcon icon={faCircleInfo} className="me-2" />
+                      Detail
+                    </Button>
+                  </Link>
+                  <Link to={`/edit/${item.id}`}>
+                    <Button className="bg-success">
+                      <FontAwesomeIcon icon={faPenToSquare} className="me-2" />
+                      Edit
+                    </Button>
+                  </Link>
+                  <Button
+                    className="bg-danger"
+                    onClick={() => handleDeleteUser(item.id)}
+                  >
+                    <FontAwesomeIcon icon={faTrashCan} className="me-2" />
+                    Delete
+                  </Button>
+                </td>
+              </tr>
+            ))}
+          </tbody>
+        </Table>
+        <Pagination className="float-end shadow-lg">
+          <PaginationItem disabled={currentPage === 1}>
+            <PaginationLink first onClick={() => handlePageChange(1)} />
           </PaginationItem>
-        ))}
-        <PaginationItem disabled={currentPage === totalPages}>
-          <PaginationLink
-            next
-            onClick={() => handlePageChange(currentPage + 1)}
-          />
-        </PaginationItem>
-        <PaginationItem disabled={currentPage === totalPages}>
-          <PaginationLink last onClick={() => handlePageChange(totalPages)} />
-        </PaginationItem>
-      </Pagination>
+          <PaginationItem disabled={currentPage === 1}>
+            <PaginationLink
+              previous
+              onClick={() => handlePageChange(currentPage - 1)}
+            />
+          </PaginationItem>
+          {[...Array(totalPages).keys()].map((page) => (
+            <PaginationItem key={page} active={currentPage === page + 1}>
+              <PaginationLink onClick={() => handlePageChange(page + 1)}>
+                {page + 1}
+              </PaginationLink>
+            </PaginationItem>
+          ))}
+          <PaginationItem disabled={currentPage === totalPages}>
+            <PaginationLink
+              next
+              onClick={() => handlePageChange(currentPage + 1)}
+            />
+          </PaginationItem>
+          <PaginationItem disabled={currentPage === totalPages}>
+            <PaginationLink last onClick={() => handlePageChange(totalPages)} />
+          </PaginationItem>
+        </Pagination>
+      </div>
     </Container>
   );
 }
